+++ src/main/java/subway/application/path/PriceLevel.java
@@ -1,4 +1,4 @@
-package subway.application;
+package subway.application.path;
 
 public enum PriceLevel {
     LEVEL1,
