+++ src/main/java/subway/application/path/PricePolicy.java
@@ -1,4 +1,4 @@
-package subway.application;
+package subway.application.path;
 
 public interface PricePolicy {
     int calculate(int distance);
