+++ src/main/java/subway/config/PriceConfiguration.java
@@ -2,8 +2,8 @@ package subway.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import subway.application.BasicPricePolicy;
-import subway.application.PricePolicy;
+import subway.application.path.BasicPricePolicy;
+import subway.application.path.PricePolicy;
 
 @Configuration
 public class PriceConfiguration {
