+++ src/main/java/subway/exception/line/DuplicateLineException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.line;
 
 public class DuplicateLineException extends RuntimeException {
