+++ src/main/java/subway/exception/line/LineNotFoundException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.line;
 
 public class LineNotFoundException extends RuntimeException {
