+++ src/main/java/subway/exception/path/IllegalPathException.java
@@ -1,6 +1,7 @@
-package subway.exception;
+package subway.exception.path;
 
 public class IllegalPathException extends RuntimeException {
+
     public IllegalPathException(String message) {
         super(message);
     }
