+++ src/main/java/subway/exception/section/IllegalDistanceException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.section;
 
 public class IllegalDistanceException extends RuntimeException {
