+++ src/main/java/subway/exception/section/IllegalSectionException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.section;
 
 public class IllegalSectionException extends RuntimeException {
