+++ src/main/java/subway/exception/station/DuplicateStationException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.station;
 
 public class DuplicateStationException extends RuntimeException {
