+++ src/main/java/subway/exception/station/StationNotFoundException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.station;
 
 public class StationNotFoundException extends RuntimeException {
