+++ src/main/java/subway/ui/LineController.java
@@ -14,8 +14,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import subway.application.LineService;
-import subway.application.SectionService;
+import subway.application.line.LineService;
+import subway.application.section.SectionService;
 import subway.dto.response.Response;
 import subway.dto.line.LineCreateRequest;
 import subway.dto.line.LineDetailResponse;
