+++ src/main/java/subway/ui/SectionController.java
@@ -11,12 +11,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import subway.application.SectionService;
+import subway.application.section.SectionService;
 import subway.dto.response.Response;
 import subway.dto.section.SectionCreateRequest;
 import subway.dto.section.SectionDeleteRequest;
 import subway.dto.section.SectionResponse;
-import subway.exception.IllegalSectionException;
+import subway.exception.section.IllegalSectionException;
 
 @RestController
 @RequestMapping("/sections")
