+++ src/test/java/subway/application/path/BasicPricePolicyTest.java
@@ -1,4 +1,4 @@
-package subway.application;
+package subway.application.path;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
